ai-services/requirements.txt
fastapi
uvicorn
python-multipart
deep-translator
SpeechRecognition
gTTS
requests
edge-tts
transformers
torch
pillow
